+++ src/crystal/tests/test_open_project.py
@@ -1,21 +1,29 @@
 from crystal import progress
 import crystal.main
-from crystal.model import Project
+from crystal.model import Project, Resource
 from crystal.progress import CancelOpenProject
+from crystal.tests.util.controls import TreeItem
 from crystal.tests.util.runner import bg_sleep
 from crystal.tests.util.server import extracted_project
 from crystal.tests.util.skip import skipTest
 from crystal.tests.util.subtests import SubtestsContext, awith_subtests
+from crystal.tests.util.wait import (
+    first_child_of_tree_item_is_not_loading_condition,
+    wait_for,
+)
 from crystal.tests.util.windows import OpenOrCreateDialog
 from crystal.util.xos import is_linux, is_mac_os, is_windows
 import os.path
+from typing import List
 from unittest import skip
 from unittest.mock import patch
 import wx
 
 
+# === Test: Start Open Project ===
+
 @skip('not yet automated: hard to automate')
-async def test_given_macos_when_double_click_crystalproj_package_in_finder_then_opens_projects() -> None:
+async def test_given_macos_when_double_click_crystalproj_package_in_finder_then_opens_project() -> None:
     pass
 
 
@@ -30,7 +38,7 @@ async def test_given_macos_when_open_crystalproj_package_in_open_dialog_then_ope
 
 
 @skip('not yet automated: hard to automate')
-async def test_given_windows_when_double_click_crystalproj_directory_in_explorer_then_opens_projects() -> None:
+async def test_given_windows_when_double_click_crystalproj_directory_in_explorer_then_opens_project() -> None:
     pass
 
 
@@ -87,6 +95,8 @@ async def test_given_linux_when_open_crystalproj_directory_and_double_click_crys
         async with ocd.open(project_dirpath, using_crystalopen=True) as mw:
             pass
 
+
+# === Test: While Opening Project ===
 
 @awith_subtests
 async def test_given_project_opening_when_click_cancel_then_returns_to_prompt_dialog(
@@ -159,3 +169,62 @@ async def test_given_project_is_corrupt_when_open_project_then_displays_error_di
                 ocd = await OpenOrCreateDialog.wait_for()
                 
                 assert did_respond_to_invalid_project_dialog
+
+
+# === Test: After Opening Project ===
+
+async def test_given_project_was_just_opened_then_no_resources_loaded_except_root_resources() -> None:
+    def loaded_resources(project: Project) -> List[Resource]:
+        # HACK: Use private API
+        resource_for_url = project._resource_for_url
+        resource_for_id = project._resource_for_id
+        assert len(resource_for_url) == len(resource_for_id)
+        return list(resource_for_url.values())
+    
+    def sort_resources(resources: List[Resource]) -> List[Resource]:
+        return sorted(resources, key=lambda r: r._id)
+    
+    with extracted_project('testdata_xkcd.crystalproj.zip') as project_dirpath:
+        async with (await OpenOrCreateDialog.wait_for()).open(project_dirpath) as mw:
+            project = Project._last_opened_project
+            assert project is not None
+            
+            # Ensure project opens with only root resources loaded
+            assert len(list(project.root_resources)) >= 1
+            resources_that_are_root = sorted(
+                [rr.resource for rr in project.root_resources],
+                key=lambda r: r._id)
+            resources_that_are_loaded = sorted(
+                loaded_resources(project),
+                key=lambda r: r._id)
+            assert resources_that_are_root == resources_that_are_loaded
+            
+            # Ensure when group expanded, only that group's members become loaded
+            if True:
+                comic_group = project.get_resource_group('Comics')
+                assert comic_group is not None
+                
+                root_ti = TreeItem.GetRootItem(mw.entity_tree.window)
+                assert root_ti is not None
+                
+                (comic_group_ti,) = [
+                    child for child in root_ti.Children
+                    if child.Text.endswith(f'- {comic_group.name}')
+                ]
+                
+                comic_group_ti.Expand()
+                await wait_for(first_child_of_tree_item_is_not_loading_condition(comic_group_ti))
+                
+                assert len(list(comic_group.members)) >= 1
+                resources_that_are_root_or_group_members = sorted(
+                    [rr.resource for rr in project.root_resources] + list(comic_group.members),
+                    key=lambda r: r._id)
+                resources_that_are_loaded = sorted(
+                    loaded_resources(project),
+                    key=lambda r: r._id)
+                assert resources_that_are_root_or_group_members == resources_that_are_loaded
+
+
+@skip('covered by: test_given_project_was_just_opened_then_no_resources_loaded_except_root_resources')
+async def test_when_resource_group_node_expanded_then_only_new_resources_loaded_are_group_members() -> None:
+    pass